requirements/common.txt
Django==1.11.6
django-model-utils==2.6
djangorestframework==3.7.1
factory_boy==2.7.0
fake-factory==0.6.0
